INCOMPATIBLE CHANGE: rework way in which we run tests
The syntax where a command could either be on a single line or
on a list is not helpful, because in YAML you can always and
easily break a long line. So, I've instead changed syntax and
now there is a single key called `command` mapping to the string
to execute. I like using keys because potentially optional keys
can be added in the future (e.g. expect failure).