Rename ./autogen to ./cmake
Now it's obvious that that is code to generate a cmake build
system. I wanted to have something more generable also capable
of generating an autotools build system, but for now I have
shelved this idead. Too much complexity will ensure. Hence this
rename to add more clarity and simplicity.