Allow to skip `tc qdisc` optional step
We need this functionality when testing performance. It seems that
mkcollector does not work when using `tc qdisc`. I'm adding this flag
so that it's possible to start opting out repositories from it.

I don't want to change behaviour tho. We'll need to add this flag to
the repos where we don't want tc qdisc. Other repositories will continue
to work (or not work) as it used to be before this flag.