cmake: add OpenSSL prebuilt for Windows
We'll be using a version of LibreSSL that we previously built
to have a working, albeit old, "OpenSSL" for Windows.

This is fine because we only use this prebuilt library for
running continuous integration tests.

Because LibreSSL/OpenSSL consists of multiple libraries, it
was necessary here to rework how prebuilt libs work.